fix: corrected IA data ordering and migrated all energy fields to BigDecimal

+++ src/main/java/com/gridmind/backend/model/EnergyBill.java
@@ -2,6 +2,8 @@ package com.gridmind.backend.model;
 import jakarta.persistence.*;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
+import java.math.BigDecimal;
+
 @Entity
 @Table(name = "energy_bills")
 public class EnergyBill {
@@ -11,8 +13,8 @@ public class EnergyBill {
     // Ruta o nombre del archivo de la foto que subirá el usuario
     private String fileUrl; 
     // Los datos que la IA logrará extraer de la foto (Lo que la compañía de luz le cobró)
-    private Double totalKwh;      // Ej. 150.5 kWh
-    private Double totalAmount;   // Ej. $45.00 dólares/pesos
+    private BigDecimal totalKwh;      // Ej. 150.5 kWh
+    private BigDecimal totalAmount;   // Ej. $45.00 dólares/pesos
     // Para saber de cuándo es la factura
     private LocalDate billingPeriodStart;
     private LocalDate billingPeriodEnd;
@@ -32,11 +34,11 @@ public class EnergyBill {
     public String getFileUrl() { return fileUrl; }
     public void setFileUrl(String fileUrl) { this.fileUrl = fileUrl; }
     
-    public Double getTotalKwh() { return totalKwh; }
-    public void setTotalKwh(Double totalKwh) { this.totalKwh = totalKwh; }
+    public BigDecimal getTotalKwh() { return totalKwh; }
+    public void setTotalKwh(BigDecimal totalKwh) { this.totalKwh = totalKwh; }
     
-    public Double getTotalAmount() { return totalAmount; }
-    public void setTotalAmount(Double totalAmount) { this.totalAmount = totalAmount; }
+    public BigDecimal getTotalAmount() { return totalAmount; }
+    public void setTotalAmount(BigDecimal totalAmount) { this.totalAmount = totalAmount; }
     
     public LocalDate getBillingPeriodStart() { return billingPeriodStart; }
     public void setBillingPeriodStart(LocalDate billingPeriodStart) { this.billingPeriodStart = billingPeriodStart; }
